requirements.txt
Django==4.0.2
python_dateutil==2.8.2
